% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-recipe.R
\name{base_recipe}
\alias{base_recipe}
\title{Default recipe blueprint for nuisance models}
\usage{
base_recipe(data, outcome, predictors)
}
\arguments{
\item{data}{Training data}

\item{outcome}{Outcome column (symbol or string)}

\item{predictors}{Character vector of predictor names}
}
\value{
a recipes::recipe
}
\description{
Default recipe blueprint for nuisance models
}
