% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-core.R
\name{dml_core_wf}
\alias{dml_core_wf}
\title{Core DML-PLR using workflows}
\usage{
dml_core_wf(
  data,
  m_wf,
  g_wf,
  folds_outer = NULL,
  n_folds = 5,
  n_rep = 1,
  vcov_type = "HC2",
  store_models = FALSE
)
}
\description{
Core DML-PLR using workflows
}
