% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-importance.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Get feature importance from nuisance models}
\usage{
get_feature_importance(dml_result, model = c("treatment", "outcome"))
}
\arguments{
\item{dml_result}{A DML result object with stored nuisance models.}

\item{model}{Which nuisance model coefficients to extract, either "treatment" or "outcome".}
}
\description{
Get feature importance from nuisance models
}
