% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load Example Data}
\usage{
load_example_data()
}
\value{
A dataframe with simulated data for DML.
}
\description{
This function loads example data for demonstrating the tiDML package.
}
\examples{
df <- tiDML::load_example_data()
head(df)
}
