% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_folds}
\alias{make_folds}
\title{Make v-fold outer folds}
\usage{
make_folds(data, n_folds = 5)
}
\arguments{
\item{data}{Data frame}

\item{n_folds}{Number of folds}
}
\value{
rsample rset
}
\description{
Make v-fold outer folds
}
