% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tune-predict.R
\name{numeric_pred}
\alias{numeric_pred}
\title{Coerce predictions to numeric vector (regression or class prob)}
\usage{
numeric_pred(fitted_wf, new_data, positive_class = NULL)
}
\description{
Coerce predictions to numeric vector (regression or class prob)
}
\keyword{internal}
