% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oob_error}
\alias{oob_error}
\title{Get treatment type: "binary_factor" or "continuous"}
\usage{
oob_error(fitted_wf)
}
\arguments{
\item{d_vec}{Treatment vector}
}
\value{
Treatment type string
}
\description{
Get treatment type: "binary_factor" or "continuous"
}
\keyword{internal}
