% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit.R
\name{oof_crossfit}
\alias{oof_crossfit}
\title{Out-of-fold cross-fitting for DML}
\usage{
oof_crossfit(data, folds, m_fit_fun, g_fit_fun, y_name, d_name)
}
\description{
Out-of-fold cross-fitting for DML
}
\keyword{internal}
