% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample.R
\name{replication_results}
\alias{replication_results}
\title{Precomputed DoubleML replication results (internal)}
\usage{
replication_results()
}
\value{
A tibble/data.frame with columns: method, seed, theta, se, lwr, upr
}
\description{
Precomputed DoubleML replication results (internal)
}
