% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-engine.R
\name{resolve_engine}
\alias{resolve_engine}
\title{Resolve engine and sensible args for a model kind}
\usage{
resolve_engine(kind, mode)
}
\description{
Resolve engine and sensible args for a model kind
}
\keyword{internal}
