% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-recipe.R
\name{resolve_recipe_factory}
\alias{resolve_recipe_factory}
\title{Resolve a recipe input into a factory function}
\usage{
resolve_recipe_factory(recipe_in = NULL)
}
\arguments{
\item{recipe_in}{NULL, a recipe object, or a factory function(data, outcome, predictors)->recipe}
}
\value{
function(data, outcome, predictors) -> recipe
}
\description{
Resolve a recipe input into a factory function
}
\keyword{internal}
