% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-engine.R
\name{resolve_spec}
\alias{resolve_spec}
\title{Accept a parsnip spec OR a short string and return a finalized spec}
\usage{
resolve_spec(model_in, mode = "regression")
}
\arguments{
\item{model_in}{a parsnip model spec or one of "rf","xgb","glmnet","linear"}

\item{mode}{"regression" or "classification"}
}
\description{
Accept a parsnip spec OR a short string and return a finalized spec
}
