% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-from-recipe.R
\name{run_dml}
\alias{run_dml}
\title{Run DML-PLR with user-supplied recipes and parsnip models}
\usage{
run_dml(
  data,
  output_recipe,
  treatment_recipe,
  output_model,
  treatment_model,
  folds_outer = NULL,
  n_folds = 5,
  vcov_type = "HC2"
)
}
\description{
Run DML-PLR with user-supplied recipes and parsnip models
}
