% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{treated_level}
\alias{treated_level}
\title{Get the "treated" level of a binary factor}
\usage{
treated_level(d_vec)
}
\arguments{
\item{d_vec}{Binary factor vector}
}
\value{
The "treated" level (second level)
}
\description{
Get the "treated" level of a binary factor
}
\keyword{internal}
