% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tune-predict.R
\name{tune_any}
\alias{tune_any}
\title{Tune any parsnip spec on given resamples with a recipe}
\usage{
tune_any(spec, rec, resamples, data_x, metrics = NULL, grid_size = 15)
}
\description{
Tune any parsnip spec on given resamples with a recipe
}
\keyword{internal}
