% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tune-predict.R
\name{crossfit_residuals}
\alias{crossfit_residuals}
\title{Cross-fit residuals across outer folds}
\usage{
crossfit_residuals(data, outcome, predictors, spec, folds_idx, recipe_factory)
}
\description{
Cross-fit residuals across outer folds
}
\keyword{internal}
