% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-rf.R
\name{dml_rf}
\alias{dml_rf}
\title{DML-PLR with Random Forest}
\usage{
dml_rf(
  data,
  y,
  d,
  x,
  folds_outer = NULL,
  v = 5,
  vcov_type = "HC2",
  trees_grid = NULL
)
}
\description{
DML-PLR with Random Forest
}
