% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_folds_stratified}
\alias{make_folds_stratified}
\title{Make stratified v-fold outer folds (recommended for binary D)}
\usage{
make_folds_stratified(data, d, n_folds = 5, n_rep = 1)
}
\arguments{
\item{data}{Data frame}

\item{d}{Treatment column name (string)}

\item{n_folds}{Number of folds}

\item{n_rep}{Number of repeats}
}
\value{
rsample rset
}
\description{
Make stratified v-fold outer folds (recommended for binary D)
}
