% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-from-recipe.R
\name{run_dml}
\alias{run_dml}
\title{Run DML-PLR with user-supplied recipes and parsnip models}
\usage{
run_dml(
  data,
  outcome_recipe,
  treatment_recipe,
  outcome_model,
  treatment_model,
  folds_outer = NULL,
  n_folds = 5,
  n_rep = 1,
  vcov_type = "HC2",
  store_models = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{outcome_recipe}{Recipe for outcome (g) model}

\item{treatment_recipe}{Recipe for treatment (m) model}

\item{outcome_model}{parsnip model spec for outcome (g) model}

\item{treatment_model}{parsnip model spec for treatment (m) model}

\item{folds_outer}{Optional rsample rset.}

\item{n_folds}{Number of outer folds when \code{folds_outer} is NULL.}

\item{n_rep}{Number of repetitions when \code{folds_outer} is NULL.}

\item{vcov_type}{Sandwich variance type (e.g., "HC2" or "HC3").}

\item{store_models}{If TRUE, keep fitted nuisance models inside the result}
}
\description{
Run DML-PLR with user-supplied recipes and parsnip models
}
